Moved test files into commands/test, fixed package
